enable tests
add dapr test with right path
