+++ test/functional/corerp/resources/daprTests/application_environment_test.go
@@ -12,15 +12,13 @@ import (
 	"github.com/project-radius/radius/test/functional/corerp"
 	"github.com/project-radius/radius/test/step"
 	"github.com/project-radius/radius/test/validation"
-
 	"github.com/stretchr/testify/require"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func Test_Application(t *testing.T) {
-	template := "testdata/corerp-resources-application.bicep"
-	name := "corerp-resources-application"
-	appNamespace := "corerp-resources-application-app"
+func Test_ApplicationAndEnvironment(t *testing.T) {
+	template := "../testdata/corerp-resources-app-env.bicep"
+	name := "corerp-resources-app-env"
 
 	test := corerp.NewCoreRPTest(t, name, []corerp.TestStep{
 		{
@@ -28,18 +26,29 @@ func Test_Application(t *testing.T) {
 			CoreRPResources: &validation.CoreRPResourceSet{
 				Resources: []validation.CoreRPResource{
 					{
-						Name: "corerp-resources-application-app",
+						Name: "corerp-resources-app-env-env",
+						Type: validation.EnvironmentsResource,
+					},
+					{
+						Name: "corerp-resources-app-env-app",
 						Type: validation.ApplicationsResource,
 					},
 				},
 			},
-			// Application should not render any K8s Objects directly
+			// Application and Environment should not render any K8s Objects directly
 			K8sObjects: &validation.K8sObjectSet{},
 			PostStepVerify: func(ctx context.Context, t *testing.T, test corerp.CoreRPTest) {
-				_, err := test.Options.K8sClient.CoreV1().Namespaces().Get(ctx, appNamespace, metav1.GetOptions{})
-				require.NoErrorf(t, err, "%s must be created", appNamespace)
+				expectedNS := []string{
+					"corerp-resources-app-env",
+					"corerp-resources-app-env-env-corerp-resources-app-env-app",
+				}
+				for _, ns := range expectedNS {
+					_, err := test.Options.K8sClient.CoreV1().Namespaces().Get(ctx, ns, metav1.GetOptions{})
+					require.NoErrorf(t, err, "%s must be created", ns)
+				}
 			},
 		},
 	})
+
 	test.Test(t)
 }
