+++ test/functional/corerp/resources/daprTests/dapr_component_name_conflict_test.go
@@ -8,30 +8,31 @@ package resource_test
 import (
 	"testing"
 
+	v1 "github.com/project-radius/radius/pkg/armrpc/api/v1"
 	"github.com/project-radius/radius/test/functional/corerp"
 	"github.com/project-radius/radius/test/step"
 	"github.com/project-radius/radius/test/validation"
 )
 
-func Test_Environment(t *testing.T) {
-	template := "testdata/corerp-resources-environment.bicep"
-	name := "corerp-resources-environment"
-
+func Test_DaprComponentNameConflict(t *testing.T) {
+	template := "../testdata/corerp-resources-dapr-component-name-conflict.bicep"
+	name := "corerp-resources-dapr-component-name-conflict"
+	t.SkipNow()
 	test := corerp.NewCoreRPTest(t, name, []corerp.TestStep{
 		{
-			Executor: step.NewDeployExecutor(template),
+			Executor: step.NewDeployErrorExecutor(template, v1.CodeInternal),
 			CoreRPResources: &validation.CoreRPResourceSet{
 				Resources: []validation.CoreRPResource{
 					{
-						Name: "corerp-resources-environment-env",
-						Type: validation.EnvironmentsResource,
+						Name: "corerp-resources-dapr-component-name-conflict",
+						Type: validation.ApplicationsResource,
 					},
 				},
 			},
-			// Environment should not render any K8s Objects directly
 			K8sObjects: &validation.K8sObjectSet{},
 		},
 	})
+	test.RequiredFeatures = []corerp.RequiredFeature{corerp.FeatureDapr}
 
 	test.Test(t)
 }
