+++ test/functional/corerp/resources/daprTests/dapr_pubsub_test.go
@@ -17,7 +17,7 @@ import (
 )
 
 func Test_DaprPubSubGeneric(t *testing.T) {
-	template := "testdata/corerp-resources-dapr-pubsub-generic.bicep"
+	template := "../testdata/corerp-resources-dapr-pubsub-generic.bicep"
 	name := "corerp-resources-dapr-pubsub-generic"
 	appNamespace := "default-corerp-resources-dapr-pubsub-generic"
 
@@ -57,7 +57,7 @@ func Test_DaprPubSubGeneric(t *testing.T) {
 }
 
 func Test_DaprPubSubServiceBus(t *testing.T) {
-	template := "testdata/corerp-resources-dapr-pubsub-servicebus.bicep"
+	template := "../testdata/corerp-resources-dapr-pubsub-servicebus.bicep"
 	name := "corerp-resources-dapr-pubsub-servicebus"
 
 	if os.Getenv("SERVICEBUS_RESOURCE_ID") == "" {
@@ -102,7 +102,7 @@ func Test_DaprPubSubServiceBus(t *testing.T) {
 }
 
 func Test_DaprPubSubServiceInvalid(t *testing.T) {
-	template := "testdata/corerp-resources-dapr-pubsub-servicebus-invalid.bicep"
+	template := "../testdata/corerp-resources-dapr-pubsub-servicebus-invalid.bicep"
 	name := "corerp-resources-dapr-pubsub-servicebus-invalid"
 
 	test := corerp.NewCoreRPTest(t, name, []corerp.TestStep{
