+++ test/functional/corerp/resources/daprTests/dapr_statestore_test.go
@@ -18,7 +18,7 @@ import (
 )
 
 func Test_DaprStateStoreGeneric(t *testing.T) {
-	template := "testdata/corerp-resources-dapr-statestore-generic.bicep"
+	template := "../testdata/corerp-resources-dapr-statestore-generic.bicep"
 	name := "corerp-resources-dapr-statestore-generic"
 	appNamespace := "default-corerp-resources-dapr-statestore-generic"
 
@@ -58,7 +58,7 @@ func Test_DaprStateStoreGeneric(t *testing.T) {
 }
 
 func Test_DaprStateStoreTableStorage(t *testing.T) {
-	template := "testdata/corerp-resources-dapr-statestore-tablestorage.bicep"
+	template := "../testdata/corerp-resources-dapr-statestore-tablestorage.bicep"
 	name := "corerp-resources-dapr-statestore-tablestorage"
 
 	if os.Getenv("TABLESTORAGE_RESOURCE_ID") == "" {
@@ -103,7 +103,7 @@ func Test_DaprStateStoreTableStorage(t *testing.T) {
 }
 
 func Test_DaprStateStore_Recipe(t *testing.T) {
-	template := "testdata/corerp-resources-dapr-statestore-recipe.bicep"
+	template := "../testdata/corerp-resources-dapr-statestore-recipe.bicep"
 	name := "corerp-resources-dss-recipe"
 	appNamespace := "corerp-resources-dss-recipe-app"
